move graph to it's own folder
