+++ extensions/graph-ui-temp/ui/dependencies-graph/dep-edge/dep-edge.tsx
