+++ extensions/graph-ui-temp/ui/dependencies-graph/dep-edge/edge.module.scss
