+++ extensions/graph-ui-temp/ui/dependencies-graph/dependencies-graph.module.scss
