+++ extensions/graph-ui-temp/ui/dependencies-graph/graph-context.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+import { ComponentWidgetSlot } from '../../graph.ui.runtime';
+
+export type ComponentGraph = {
+  componentWidgets: ComponentWidgetSlot;
+};
+
+export const ComponentGraphContext = createContext<ComponentGraph | undefined>(undefined);
