+++ extensions/graph-ui-temp/ui/dependencies-graph/index.ts
@@ -1,0 +1,3 @@
+export { DependenciesGraph } from './dependencies-graph';
+export { ComponentGraphContext, ComponentGraph } from './graph-context';
+export type { DependenciesGraphProps } from './dependencies-graph';
