+++ extensions/graph-ui-temp/ui/dependencies-graph/minimap.ts
